+++ AssetsFlow.Client.React/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 - Users Service, Demo user
 - Update frontend to store and send JWT tokens + auth api
-- Unified gateway - auth + fetch trades
+- Unified gateway - auth + fetch trades steps
 
 ## [0.9.019-c] - 2025-06-24
 ### Added